LayoutTests/storage/indexeddb/modern/transaction-scheduler-3-expected.txt

This test makes sure that a write transaction is blocked by a read transaction with overlapping scope. It also makes sure the write transaction starts after the read transaction completes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


ALERT: Upgrade needed: Old version - 0 New version - 1
ALERT: versionchange transaction completed
ALERT: Success opening database connection - Starting readonly transaction
ALERT: Creating write transaction
ALERT: Read transaction complete - [object Event]
ALERT: Write transaction put success
ALERT: Write transaction complete - [object Event]
PASS successfullyParsed is true

TEST COMPLETE
